MaartEngine Release V0.1
The first release of the 2D engine MaartEngine it is a basic engine with a simple game so you can find out how to use all the functions. it is a Visual Studio project. you vave to import it in visual studio and the edit the DemoGame.cs FIle. that is the file where you can put your game source code in. you can also add more functions bvecause all the source code is in the project.